Revert moving patch from init to test
